Initial code for component sample Owner
The component sample for Owner contains the following updates as
compared to http-server-to2-sample and http-client-to0-sample:
- The To0-Client and To2-Server have been merged together. The combined
Owner sample starts the TO2 service and executes the To0-Client on a
separate thread every 300 seconds.
- The application can be configured using properties.
- Support for PKCS11 keystore type has been added. The owner keys can
now be read from PKCS11 keystore.
- A top-level 'demo' directory is created for owner that contains
Docker setup that cofigures JVM and installs softHSM as the PKCS11
keystore.

TO-DOs:
- Update README

Signed-off-by: Chandrakar, Prateek <prateek.chandrakar@intel.com>